refactor feedback.service.js and sendgrid-service.js to break curcular dependency and add test for function sendSnapshotSubmissionToDRO
